+++ start.sh
@@ -5,7 +5,7 @@ usage() {
     exit
 }
 
-. .release/base.sh || exit
+. .release/base.sh || exit 1
 
 find_new_version() {
     if ! git branch --show-current | grep -q ^release;
@@ -35,8 +35,7 @@ find_new_version() {
 }
 
 write_template() {
-    cat << EO_CHANGE >> .release/new.txt
-
+    cat <<EO_CHANGE > .release/new.txt
 
 ### [$NEW_VERSION] - $YMD
 
@@ -59,11 +58,11 @@ add_commit_messages() {
     LAST_TAG=$(git describe --tags --abbrev=0)
     if [ "$LAST_TAG" != "" ]; then
         # append log entries since the last tag (release)
-        git log --pretty=format:"- %s" "$LAST_TAG..HEAD" >> .release/new.txt
+        git log --pretty=format:"- %s%n%b" "$LAST_TAG..HEAD" >> .release/new.txt
     fi
 }
 
-add_release_link() {
+changelog_append_release_link() {
 
     if grep -q "^\[$NEW_VERSION\]:" "$CHANGELOG"; then
         echo "CHANGELOG URL for $NEW_VERSION exists"
@@ -75,8 +74,25 @@ add_release_link() {
     fi
 }
 
-update_changes() {
-    # insert contents of new.txt into CHANGELOG.md after marker
+changelog_add_header()
+{
+    if ! grep -qi '# Changelog' "$CHANGELOG"; then
+        echo "inserting: # Changelog"
+        sed -i '' \
+            -e '1s|^|# Changelog\n\nThe format is based on [Keep a Changelog](https://keepachangelog.com/).\n\n|' \
+            "$CHANGELOG"
+    fi
+
+    if ! grep -q '# Unreleased' "$CHANGELOG"; then
+        echo "inserting: ### Unreleased"
+        sed -i '' \
+            -e '1,/##/ s/##/### Unreleased\n\n##/' \
+            "$CHANGELOG"
+    fi
+}
+
+changelog_add_release_template() {
+
     if grep -qE "^#* $NEW_VERSION|^#* \[$NEW_VERSION\]" "$CHANGELOG"; then
         echo "CHANGELOG entry for $NEW_VERSION exists"
     else
@@ -84,40 +100,81 @@ update_changes() {
 		add_commit_messages
 		echo "" >> .release/new.txt
 
-        if head "$CHANGELOG" | grep -q Unreleased;
-            sed -i '' -e "/### Unreleased$/r .release/new.txt" "$CHANGELOG"
-        then
-            sed -i '' -e "/#### N.N.N.*$/r .release/new.txt" "$CHANGELOG"
-        fi
+        # insert contents of new.txt into CHANGELOG.md after marker
+        sed -i '' -e "/## Unreleased$/r .release/new.txt" "$CHANGELOG"
         rm .release/new.txt
     fi
 
-    add_release_link
+    changelog_append_release_link
     if command -v open; then open "$CHANGELOG"; fi
 
     echo
     echo "AFTER editing $CHANGELOG, run: .release/submit.sh"
 }
 
+changelog_check_tag_urls()
+{
+    echo "checking tag URLs..."
+    git fetch --tags
+    local REPO_URL; REPO_URL="$(gh repo view --json url -q '.url')"
+
+    for _tag in $(git tag); do
+        local _ver="${_tag//v}"
+        local _ver_uri="[$_ver]: $REPO_URL/releases/tag/$_tag"
+
+        if ! grep -Fq "$_ver_uri" "$CHANGELOG"; then
+            if ! grep -Fq "[$_ver]:" "$CHANGELOG"; then
+                echo "$_ver_uri" >> "$CHANGELOG"
+            else
+                echo "INVALID URI in CHANGELOG"
+                grep -F "[$_ver]:" "$CHANGELOG"
+                echo " ----- should be ------"
+                echo "$_ver_uri"
+                echo
+            fi
+        fi
+    done
+    echo
+}
+
 constrain_publish() {
-    # many modules have a .npmignore (one more file) to reduce/limit what
-    # gets published. Instead...
-    if [ -f .npmignore ]; then
-        echo "CONSIDER: instead of maintaining .npmignore, add the much shorter list of"
-        echo "          files that should be published to [files] in package.json."
+    local _main; _main=$(node -e 'console.log(require("./package.json").main)')
+    if [ "$_main" = "undefined" ]; then
+        echo "CONSIDER: package.json has no [main] section. You can likely reduce"
+        echo "          the published package size by populating it."
         echo
         echo "   https://docs.npmjs.com/cli/v10/configuring-npm/package-json#files"
         echo
-        echo "the current contents of [files] in package.json:"
-        echo
-        node -e 'console.log(require("./package.json").files)'
     fi
 
-    # local _main; _main=$(node -e 'console.log(require("./package.json").main)')
-    # if [ "$_main" = "undefined" ] && [ ! -f index.js ]; then
-    #     echo ""
-    # fi
+    # many modules have a .npmignore (one more file) to reduce/limit what
+    # gets published.
+    if [ -f .npmignore ]; then
+        echo "CONSIDER: instead of maintaining .npmignore, add the much shorter list of"
+        echo "          files that should be published to [files] in package.json. -^"
+        echo
+    fi
 }
+
+self_update()
+{
+    (
+        cd .release || exit
+
+        if [ "$(git branch --show-current)" != "main" ]; then
+            git checkout main
+        fi
+
+        _pull=$(git pull origin main)
+    )
+
+    if [ "$_pull" != "Already up to date." ]; then
+        git add .release
+        . .release/base.sh
+    fi
+}
+
+self_update
 
 find_new_version "$@"
 
@@ -133,11 +190,10 @@ if branch_is_main; then
 fi
 
 find_changelog
-update_changes
+changelog_add_header
+changelog_add_release_template
+changelog_check_tag_urls
 constrain_publish
 
 git add package.json
 git add "$CHANGELOG"
-
-# update .release submodule, but leave it to author to review/check in
-cd .release && git checkout main && git pull origin main && cd ..